requirements.txt
torch
transformers
pillow
opencv-python
streamlit  # Or flask, depending on the choice
